[US9] Changed all prices to Long

+++ backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/entity/TicketType.java
@@ -30,7 +30,7 @@ public class TicketType {
     private String title;
 
     @Column
-    private double price;
+    private Long price;
 
     @ManyToOne(fetch = FetchType.EAGER, cascade = CascadeType.ALL)
     @JoinColumn(name = "SECTOR_ID", nullable = true)
